[4.0] Popup Messages/Alerts
In J4 we currently have two types of alert. One uses a custom element and is a "popup" the other is a regular div or "inline". This PR is to improve the accessibility of the popup messages/alerts.

Currently we make use of color which conveys no meaning for a color blind user. Accessibility golden rule is to never rely on just one visual indicator so this PR adds an icon. This also makes the popup alerts consisten with the inline alerts updated #25116

Finally the title of the popup used an H4. @zwiastunsw requested here https://github.com/joomla-projects/custom-elements/issues/99#issuecomment-423668924 that this was not good for accessibility as the headings may not be nested correctly. So I have changed the H4 to a class=alert-heading (I have not customised the style of that class as it would be a waste of time as the admin template would change it) It can always be styled later

This is based on the discussions joomla-projects/custom-elements#99