[4.0] Remove unused file
As far as we can tell this file is never used and never has been. It did not exist in J3 and can only assume it was a copy paste error when the component was updated for J4. The strings used are not in any language file etc.

PR for #25125