Add modified date of a master item to the associations table
This adds a new column "assocParams" to the association table where the modified date of the master item is stored. The column for the master item always gets its latest modified date, its children get the modified date when the association is created.

The table of menu items does not yet have a modified date, so the column remains empty.

Also, a new AssociationsHelperClass is added for all Masterlanguage matters.