fix display of item association tooltips
Since the tooltips are hidden by default and are only displayed when the previous sibling element is focused in the DOM, tooltips that have the same role and are applied as popovers using bootstrap can no longer be displayed because they do not have a previous focusable sibling element. They differ in that they get the class '.show' when activated.

So, to fix the problem, the css statements do not apply to tooltips with the role that have the class 'show'.