A fix to allow "0" (zero as a string) to be an "option" element text
Currently there is no way to have an option element of a select with string "0" as a text, as it evaluates to false and getting replaced with value.

Real life example: there are kids clothing sizes "00", "0', "2", "4", "6" etc., and when we have an array with indexes not matching these values e.g. [25 => '00', 26 => '0', 27 => '2', 28 => '4', 29 => '6'] then after creating select options via JHtml::_('select.option', ...) we end up with a dropdown select with [00, 26, 2, 4, 6] instead of [00, 0, 2, 4, 6] choices.