[4.0] Inline Messages/Alerts (#25116)
* [4.0] Inline Messages/Alerts [a11y]

In J4 we currently have two types of alert. One uses a custom element and is a "popup" the other is a regular div or "inline". This PR is to improve the accessibility of the inline messages/alerts.

Currently we only convey the importance of the text by use of color which conveys no meaning for a color blind user. Accessibility golden rule is to never rely on just one visual indicator so this PR adds an icon.

Secondly we need to convey the importance of the text to screen readers. The text alone is not always sufficient to do that so this PR adds a sr-only text

* tab

* tabs

* Update administrator/components/com_contact/tmpl/contacts/default.php

Co-Authored-By: Quy <quy@fluxbb.org>

* thanks @quy