[4.0] com_users multiple groups
In the list of users if a user is in more than 2 groups then it just displays "Multiple groups" and then a tooltip listing the groups.

This PR changes it from a completely inaccessible tooltip to the new div approach and instead of putting the groups on new lines it correctly puts them in a list. Also adds tabindex so that keyboard users can access this information.