[4.0] Character count for meta description
A request was made to have a character count when entering a meta description. As a Proof of Concept I have done this and applied it to the Site Meta Description in the global configuration as shown in the screenshot. (If accepted it is quick to add it to all the meta description fields

As this uses a new field layout "charcount" it can be applied to any textarea field by adding `layout="joomla.form.field.charcount"` to the field definition in the xml.

The character count has a default maximum of 200 characters but this can be configured in the field xml by adding `maxlength="160"`

Accessibility is obviously very important and this PR addresses this as well. The meta description text is announced when the user focuses on the input together with the string to indicate the characters used. Obviously a live count announced by the screenreader would be distracting so it waits 2000 milliseconds before updating the user with the screenreader user with the current count.

The string `86 characters remaining of 160 characters.` is a regular joomla language string except the numbers are replaced  by the script. In this example I chose
`"{remaining} characters remaining of {maxlength} characters."`. Available placeholders available are {remaining}, {maxlength}, {length}