update outdated child of a master item
This adds a new edit view for updating a child item when it is oudated with the master item. There it is just possible to save the target and update his masters modified date.
When versions are enabled for this itemtype then there is a compare view displayed for the reference/master otherwise the edit view is displayed.