[4.0] Role = switch
After extensive testing last night it became apparent that the role=switch is not correct and not needed for this code. It is not well supported and we don't need it. This pr removes it from the swticher buttons. You will only notice a change if you are using a screenreader where the content is announced more correctly