[4.0] Permissions accessibility
As a result of the accessibility performed last night on the Permissions page it was realised that it would be best for screen reader users to know what the changed setting is. For example if they chose "inherited" they need to know that this means that "allowed - inherited" has been selected.

The super easy way to do this is to wrap the "allowed - inherited" span in the html element output. This way whenever the output is updated it is automatically announced by the screen reader.

Technically we could remove the span and put the class on the `<output>` element but as that would require changing the javascript as well I didn;t see the need to do that. The code in this PR works perfectly without making that change