[4.0] Drag and Drop styling
When you try to do a drag and drop reordering there is supposed to be a color applied to the row as you drag it. Due to a bug when @variables support was added to the scss file this wasnt working because there were TWO colours.

This pr fixes that. The choice of colour is up to the template and is beyond the scope of this PR