[4.0] Touch keyboard support
Unlike with a physical keyboard a mobile or touch device has multiple different keyboards. This PR ensures that joomla uses the correct mobile keyboard which greatly increases usability on mobile/touch. Although support is not 100% on desktop touchscreen browsers it is really mainly for devices with no physical keyboards

This PR adds an attribute as appropriate  `inputmode="xxxxxx"` for the field types url, tel, number and email.

This can be tested on any mobile/touch device.