[4.0] Permission rules - accessibility etc (#25366)
We have a large block of text on every permissions tab right at the bottom. Among the many problems with a large block of text like this is accessibility. Text like "see notes at the bottom" just doesn't work for assistive tech. Luckily there is a really easy solution. Using the html elements `<summary>` and `<details>`. The `<details>` element generates a simple no-JavaScript widget to show/hide element contents, optionally by clicking on its child `<summary>` element. It is fully accessible as the summary element is presented as a button to the accessibility tree.

Note these elements are not supported in IE11 (who cares) or EDGE (pre chromium) but by the time J4 will be released it wont be an issue ;)

### Testing instructions
npm i is needed "only" to check my bad scss ;) You can still test with patchtester you just wont get all the styling