[4.0] Featured icon display
If you have a user who does **not** have core.edit.state permissions for com_content.article then the icon displayed for "featured/unfeatured" looks wrong. The problem is that the style comes from the `tbody-icon` on the button and the code in the joomla.icons.action-button layout has two bugs.

The first bug is that **disabled** was in the class attribute - it should be its own attribute. This meant that the button was never disabled. Which is probably why the second bug was introduced - the entire `if(!empty($disabled))` test is not required and because it is not wrapped in a button it never gets the `tbody-icon` class and looks **wrong**

### Testing instructions
Create a new user at the manager level. In article options change the permissions for manager so that **Edit State** is disabled.

Login as that user and go to the list of articles.

### Before PR

### After PR