[4.0] Text Filters - accessibility etc (#25394)
We have a large block of text on the text filter tab at the top and at the bottom. Among the many problems with a large block of text like this is accessibility. Text like "see notes at the bottom" just doesn't work for assistive tech. Luckily there is a really easy solution. Using the html elements `<summary>` and `<details>`. The `<details>` element generates a simple no-JavaScript widget to show/hide element contents, optionally by clicking on its child `<summary>` element. It is fully accessible as the summary element is presented as a button to the accessibility tree.