[4.0] Sortable icon and title text
This PR fixes the bug that the burger icon should be grey when sorting is disabled. (you will need to npm to build the css to test this part)

This PR updates the markup for the title text. It removes a class that was for mootools and as its just a regular title we dont need to pass the text through a helper.