[4.0] alerts accessibility (#25421)
When changing the permission of something if there is a child group then a popup notice is displayed. For a sighted user they can still carry on and change the permissions of other properties. However for a user using a screen reader that is not the case. After changing a permission the focus then moves to the popup notice. This has to be closed before the user can continue but the focus is returned to the url bar and they then have to navigate through the entire page again.

the same is true for all alerts although it is particularly bad with this one.

The fix is in two parts.
Part one is to stop moving the focus to the popup notice and this change is in the custom element and can be found here https://github.com/joomla-projects/custom-elements/pull/120

This is part two
This adds the attribute aria-live to the message container. This will ensure that the message will be announced by the screenreader even though we don't give it focus. Note it must be on the message container because the aria-live must be present in the document on page load - even if there are no messages - for aria-live to work.

aria-live=polite means that the message will be announced next ie after what is being announced now
there is another option which is not appropriate here
aria-live=assertive which means that the message will interrupt anything currently being announced.