[4.0] Font preview - not a form
Unless I am missing something, which is quite possible, then the font preview in com_templates is not a form and this PR removes the form related code from the preview.

To test this you will need to copy a font to your template folder. Once you have done that you should see the following preview when you click on the font