[4.0] Menu description
We have a field called Menu Description which as the name implies is used to provide an optional description of the menu. But we only display it on the menu edit form.

This field really is functionally the same as the "note" field that we have elsewhere which is displayed on the list views.

So this simple PR adds the display of the description below the menu name