[4.0] Rules - tooltips
Although its never used in core it is documented that each access rule can have a description.

This PR makes sure that it is the new style accessible tip.

To test you will need to edit a components access.xml and add  `description="something here"` to one of the rules