[4.0] Installation - create a username
For historical reasons that I bet no one can remember Joomla doesn't currently let you enter the real name of the Super User when installing Joomla. As a result every site on installation has a user with the name "Super User". This regularly causes issues as it gets confused with the usergroup "Super Users". That might be a translation issue or simply because there is just one character difference I don't know and it doesnt really matter.

As today there is no valid reason to force the name to be "Super User" and this PR prompts for the real name as well as the username during installation.

### To test.
Apply this PR using patchtester.
delete configuration.php
re-install joomla

(Please no bikeshedding about other issues in the installer)