Tree sorting in templates file tree
This is a weird one and I cant say I fully understand it but this resolves the problem described below.

If you have a large number of files in a folder then when you look at the list of files in the tree you will see that they are not in alphabetical order. They are in alphabetical order "in groups" which is why you need more than usual files to see it.

To test copy a folder with a lot files to `templates\protostar\html`

I used this folder as a source `administrator\components\com_templates\views\template\tmpl`

Now when you view the file tree in the template component you will see the sirting is wrong as per screenshot

Apply this PR and retest and the sirting is as expected.