[4.0] Dashboard admin submenu modules
The markup for these modules is incorrect. The h2 heading should only contain the title not the title plus dropdown menu.

In particular this causes issues for assistive technology as it doesn't understand what the title is. You can test this easily using the skip-to menu (hit tab) and you will see that the h2 has been truncated to nothing.

Apply this PR and its all good