fix bug for correct parent_date when adding child from other association
When adding a child from another association, the parent_date for this child wasn't updated to the parent_date of the new association. This is fixed now. The key of each item before saving this association is compared with each other, in order to be able to check, if the item comes from another association and to update then the parent_date.