[4.0] Remove menuitem letter-spacing
Really not a fan of changing the letter spacing on a font. Especially as it can cause undesired issues with non latin characters or when the user has _zoomed_ on the page to increase the text size for accessibility. The font designers are pretty good and know more about typography than we do.