[4.0] Enable com_finder (#25839)
Prior to Joomla 4 we had two search options - we now just have finder aka smart search. So it makes no sense for this to be disabled on NEW installs. This PR enables the plugin on NEW installs so that you know longer get the warning etc.

This does not change EXISTING sites