[4.0] Permissions widths
There is no specific width set on the columns in the permissions table. As a result as you move between access levels you will see the columns jump around depending on the width of the content on the first row in the last column.

This PR resolves this by setting explicit widths for the columns