Fix issue #24782:
This basically reverts the change made to this line in commit c778973, but that is a merge commit
and none of the merged parents have those change, and the commit does not give an explanation why
this change is needed. It is not needed at all from my point of view, so reverting it here.

This fixes #24782 and should not have any side effects, as the CategoryService is being instanciated
with a correct context (the component), so passing the $section is not needed. We want all categories.