[4.0] Table border
This simple PR adds a border to all tables. It is the same css as the border added to cards on the dashboard and to fieldsets in a component options.

It adds some extra consistency to the UI and addresses the problem that the last row in a table merges into the background on hover.