[4.0] Singular strings
PR for #26284

### Preamble
This is referring to the english values in the strings **not** any translation
This is **not** about changing any language constants
This _is_ about consistency in the english used (if it was any other language it would not even be discussed as we would not know about it)

### Issue
We have  at least **THREE** different ways that singular strings are written in english

> COM_REDIRECT_N_ITEMS_ARCHIVED="%d links archived."
COM_REDIRECT_N_ITEMS_ARCHIVED_1="Link archived."

> COM_CONTACT_N_ITEMS_ARCHIVED="%d contacts archived."
COM_CONTACT_N_ITEMS_ARCHIVED_1="%d contact archived."

> COM_USERS_USERS_N_ITEMS_DELETED="%d users deleted."
COM_USERS_USERS_N_ITEMS_DELETED_1="1 user deleted."

Language strings should be consistent in style and language. This is not only important for consistency and creating a professional impression it is also import for accessibility.

> providing consistent labels for the repeated components reduces the cognitive load of users with cognitive disabilities

### Changes in this PR
All the **values** have been updated to one single format and the enGB style guide updated so that singular strings are

> COM_REDIRECT_N_ITEMS_ARCHIVED="%d links archived."
COM_REDIRECT_N_ITEMS_ARCHIVED_1="Link archived."

> COM_CONTACT_N_ITEMS_ARCHIVED="%d contacts archived."
COM_CONTACT_N_ITEMS_ARCHIVED_1="Contact archived."

> COM_USERS_USERS_N_ITEMS_DELETED="%d users deleted."
COM_USERS_USERS_N_ITEMS_DELETED_1="User deleted."

#### Note
This is not a global search and replace. There are some string were I have kept the original text as it makes sense for that specific message