[4.0] Autocomplete fields
Pull request for #26459

Unless I am missing something then this docblock in all fields is wrong
https://github.com/joomla/joomla-cms/blob/bc42bb1e43d96334c38222dbe49b83f898413f2a/layouts/joomla/form/field/password.php#L20

All we ever do is this
https://github.com/joomla/joomla-cms/blob/bc42bb1e43d96334c38222dbe49b83f898413f2a/layouts/joomla/form/field/password.php#L72

Which means that we can never have any value for autocomplete in the xml other than 'off'

You can see this by changing the value for autocomplete here to anything else and instead of the autocomplete being output with the new value it is removed.

https://github.com/joomla/joomla-cms/blob/bc42bb1e43d96334c38222dbe49b83f898413f2a/administrator/components/com_users/forms/user.xml#L20-L30

Autocomplete has a lot of useful values that we could/should be using see https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete

For example `autocomplete=off` really doesn't do anything on password fields. Instead we should be using `autocomplete="new-password"`

It looks to me that it can be _fixed_ by changing the code to
`$autocomplete ? '' : 'autocomplete="' . $autocomplete .'"',`

And autocomplete is not the same as autofill
> Chrome will still respect this tag for autocomplete data, it will not respect it for autofill data

But I am probably missing something
cc @HLeithner @wilsonge @mbabker