[4.0] Mobile template toolbar
This PR contains the markup for a proposed change to the display of the toggled toolbar.

It is does NOT contain the css changes that are required. I really could do with some help with converting the css below to appropriate scss - or even better scss

You can paste the css below into your browser dev tools after applying the PR to see the intended behaviour

Pull Requests welcome to create the scss

@media (max-width: 575.98px) {
  .toggler-toolbar {
      top: 20px;
  }

  .toggler-toolbar.collapsed .toggler-toolbar-icon::before {
      font: normal normal normal 30px/1 'Font Awesome 5 Free';
      content: "\f085";
      color: var(--toggle-color);
  }

  .toggler-toolbar .toggler-toolbar-icon::before {
      font: normal normal normal 30px/1 'Font Awesome 5 Free';
      content: "\f00d";
      color: var(--toggle-color);
  }
  .subhead joomla-toolbar-button, .subhead .btn-group, .subhead .btn {
      width: 100%;
  }
  .subhead .btn [class^='icon-'], .subhead .btn [class*=' icon-'], .subhead .btn [class^='fa-'], .subhead .btn [class*=' fa-'] {
    float: left;
  }
}