[4.0] logout module profile link
the class unstyled is an old b2 class not present in cassiopeia or bs4. replaced with list-unstyled so it has the desired impact

### testing
enable the show profile link in the site login module. Login and see the link above the logout button
it should be a link and not a bullet link