[4.0] Article Options
There is an option for com_content to hide the article options tab for the editing layout of individual articles - I'll be honest and say I never knew it was there.

Unfortunately it didn't have any effect.

This PR fixes that

### Testing

Set Show Article Options to hide

### Expected behaviour with this PR
The Options tab is not present any more