[4.0] Configure content edit tabs
There are a lot of tabs and a lot of options when editing content. We already have the ability to hide come of those tabs. This PR simply extends that to all the additional tabs. This PR does not change the default display it just extends the existing ability to hide ssome tabs to all the tabs.