[4.0] Display Menu Title
Sometimes you dont want to display the menu title - for example if you have an image instead.

The wording for this option is "Add Menu Title"

I find this confusing as the default value is yes and it would be the expected behaviour etc for a menu title for it to be displayed.

This PR proposes changing the text to "Display Menu Title"