License Key implementation in Joomla 4 has a confusing end user UX
gh-26583

Plugin to warn users of any Download Keys they need to enter.

When no extension requires a Download Key to be entered the plugin is
not even shown so as not to confuse users (it would link to a page with
an empty listing which makes no sense for **most** Joomla! users).