[4.0] Templates Preview image
When you click on the thumbnail image of a template you see a larger image of the template. The current title of that popup is "Template Preview". This is not the correct string as you might have on the same row a message saying "Preview not available". In addition the value of the string being incorrect it was also using the same key as the preview button which does something different - so might cause translation issues.

This PR introduces a new string to use for the popup which correctly iddentifies the image as a screenshot and includes the name of the template