[4.0] Tag field description
In two places the tags field was still displaying a useless description.

Category Blog menu item
Articles Category module

This simple PR removes the display of that string. The string is kept as I have tried not to remove strings from any of the generic language files as they may be used by non-core extensions