- Fixed "answer provided by user" issue.
Worked on "onCheckAnswer" function. The "$code" parameter's value wasn't implementing inside the function. While you call "onCheckAnswer" function, it used to take "response" from the form which was conflicting if multiple reCAPTCHA found in the respective page.

Now (on this fix): if anyone sends any value to "$code" parameter on "onCheckAnswer" function, the "$code" parameter will take "$code" as "response".