[4.0] Download Keys quick icon could crash the site (#27020)
If there is a leftover update site from an extension you've removed or, more generally, the extension's XML manifest is no longer present the site breaks because of the Download Keys quick icon plugin. 

The problem is that InstallerHelper::getInstallationXML() calls simplexml_load_file and always expect a SimpleXMLElement object. On load failure the simplexml_load_file function returns false.

Fixed by changing the method signature to expect SimpleXMLElement or null and converting a false result into null. The rest of the core works as-is.