Square bracket encoding in pagination links
I introduced the method "queryEncode" to encode square brackets in pagination links.
AbstractURI->toString() returns unencoded brackets, so I had to find a workaround, but I'm not sure it's the cleanest way to do this.
The method is defined as protected in Joomla\CMS\MVC\View\JsonApiView so that it can be used by other extending classes, like com_config api views.
Fixed an error in com_config component view pagination (L109).