[4.0] Move Private Messages to Users
As the component menu is now used for components that create some form of site content this Pr moves the private messages component from the Component menu to the users menu.

When testing simply applying the patch will only add the messages menu items to the users menu. It will not remove it from the component menu. As the sql change is data and not structure you cannot use the database fix either. It will have to be a direct update in the sql using phpmyadmin or similar