[4.0] Select Client/Administrator filter [a11y]
The labels for some of these fields was missing or not being used due to the layout. In addition the labels have been standardised to "Select Client" which is more in keeping with other similar labels.

To test you will need to apply the patch and then view the source code to see that all these filters have an sr-only label of Select Client