Update News Need ID in Single News Feed menu item
At present the News Feed ID mentioned/chosen in the SQL of the menu item with title Single News Feed is 4.  Whereas there are only three News Feed Items defined in the Sample Data Learn and their IDs are 1, 2, and 3.  So, there is no News Feed item with ID 4.  This results in error when the menu item Single News Feed is chosen in the front end when the Sample Data Learn is installed.