[4.0] Accessibility on associations
On a multilingual site with more than two languages the language indicators have a div for displaying the _tooltip_

The way the id for the div was generated it might not be unique. This PR ensures that it is unique.

Applied to menus, categories, content, contacts, newsfeeds (don't think there is anywhere else?)