[4.0] Search inputmode (#27426)
Mobile and touch devices will load a different keyboard depending on the inputmode. This is already implemented for several fields eg url and email

This PR enables it for search fields

As search fields are not a separate field type but type="text" we either have to create a new field type or it has to be added via the form xml. Support for the xml option  was (surprisingly) already implemented so I just had to add the inputmode=search to the xml.

To test this you will need either a mobile device or a touch enabled computer with a virtual keyboard.

When you enter the search field you will see a slightly different keyboard than normal - the actual layout of which is dependent on your device. On my touch screen it looks like this