[4.0] Admin custom menu item
When you create a custom admin menu there is the possibility to add translations for each menu item using the language override system.

The actual text that will be displayed for the menu item is displayed in a read only field.

This PR gives it some styling love and care