[4.0] Filter by parent category
On a complex site with a lot of categories it can be very difficult to sort the categories using drag and drop because you cannot get the relevant tree on the same page.

This adds a new filter to com_category that allows you to select the parent category. It's particularly useful when used with the max levels filter.

Testing is easiest if you have the test sample data installed as that has a lot of categories