[4.0] TinyMCE element_path
Tinymce has an option in the plugin called Element Path but it doesn't do anything as there is no code to do anything.

With this PR you can no disable/enable the display of the current element in the bottom left hand corner

### Testing
Make sure that you set the option to be enabled/disabled in the correct set. Default for a super user is set 0

### Additional Note
This is also broken in Joomla 3