[4.0] TinyMCE context menu
Tinymce has an option in the plugin called Context Menu but it doesn't do anything as there is no code to do anything.

With this PR you can no disable/enable the ability to right click on an item in the editor and get a contextual menu. When this is disabled you get the context menu from your browser

### Testing
Make sure that you set the option to be enabled/disabled in the correct set. Default for a super user is set 0

### Additional Note
This is also broken in Joomla 3