[4.0] gray-600 Cassiopeia template
I have no idea why the value for gray-600 is different to the default bootstrap value but as a result it is failing the contrast checks.

You can see this in action on the article info fields.

This PR changes the value back to the bootstrap default which passes the accessibility contrast check and its almost impossible to visually differentiate

For testing don't forget to npm -i