[4.0] Toggle menu selection template style
If you go to the template style and try to select which menu items etc that the style will be applied to then the wording is "Toggle selection".

The reality is that it does not toggle the selection but does either "select all" or "select none".

This PR corrects that so the toggle really does toggle the state of the checkboxes.

(Although the code is different in j3 the problem exists there as well so I can backport it if required @hleithner @zero-24)

The same problem with the toggle not being a toggle exists in other places but again as the code is different in each place I will address those in their own PR.

To test don't forget to do `npm i` or `node build.js --compile-js`