[4.0] Toolbar - Options class
The code that adds the class "ml-auto" to the Options button on all the toolbars has an extra space. It doesn't break anything but technically its invalid markup.

To test view the source of any Options button in any component before and after this PR.

It will change from
<joomla-toolbar-button class=" ml-auto">
to
<joomla-toolbar-button class="ml-auto">