[4.0] superfluous markup - backtrace
This PR removes the class `Table` and `TD` from the markup in the error backtrace. Code review only needed as the class doesn't exist in the css. Or enable debug mode and then in the frontend force an error eg go to `index.php?option=com_wilson`

other than the removal of the classes that will be seen in the source there will be no other change