[4.0] Language module
When the language switcher is set to display flags a title is set for each image with the full language name.

But if a language is set not to have a flag then we get the SEF value instead of the flag but there is no title.

This simple PR adds the title in that situation.

To test
Create a multilingual site
Open one of the content languages and set the image value to none
Make sure the language switcher module is set to use image flags
Hover over the images and the text in the language switcher module in the front end
After this PR all languages will display the title