[4.0] Language Mail Templates
When a content language has a flag a title is set for each image with the full language name.

But if a language is set not to have a flag then we get the SEF value instead of the flag but there is no title.

This simple PR adds the title in that situation.

To test
Create a multilingual site
Open one of the content languages and set the image value to none
go to the system menu and he mail templates
Hover over the images

After this PR all languages will display the title