[4.0] Filtered by - accessibility
In a previous pull request I added code which adds a caption to every table AND indicates how the table is sorted.

To improve this further we should also include how the table is filtered.

This DRAFT pr which I hope the javascript gurus can help with adds the filtered information.

I have only enabled this for the table of contacts at this time while in draft

Currently the output is like this
<caption id="captionTable" class="sr-only">
Table of Contacts
<span id="orderedBy">Sorted by:Title - descending</span>
<span id="filteredBy">Filtered by:Public3</span>
</caption>

As you can see the filtered by is displaying just the selected value of the selected filters, which is kind of ok when its text but is meaningless when it is a number.
Ideally it would say
Filtered by: Access level = Public, Max Levels =3