[4.0] Dashboards
Before this PR In the dashboards of joomla 4 we have a situation where sometimes there is a large space between _modules_

This is because the _modules_ generated by the system are in a different div than that of the more traditional modules.

By moving them all into the same div that problem disappears.