[4.0] Dashboard permissions
The component dashboard does not "hide" components that a user does not have access to.

This PR removes the component from the dashboard if the user doesn't have access to that component.

### To test
1. create a new user of level admin
2. Change the permissions for com_banners so that the admin userlevel is denied access to the Administration Interface
3. Log in as that user

### Expected result
banners no longer present on the menu AND banners no longer present on the dashboard

### Actual Result
banners still present on the dashboard with a 403 denied if you try to click

### Apply the PR
banners is not present on the dashboard