[4.0] Codemirror update
Codemirror was locked at a specific version that is now very old.

This PR changes the setting in package.json so that it follows semver

Codemirror has always been very good at following semver so this should be totally safe.

(the update process also updated tinymce at the same time)