[4.0] User Menu heaader
The display of the user name in the user dropdown menu looks a little off due to lack of styling.

This add an icon before the username and the text "signed in as " as a prefix. (The text is as used on github)

There are css changes so npm i or node buuld.js --compile-css