[4.0] TinyMCE new features
With the release of tinymcec 5.1 we have some extra features available. This PR adds the [toolbar drawer](https://www.tiny.cloud/docs/configure/editor-appearance/#toolbar_drawer) and the [sticky toolbar](https://www.tiny.cloud/docs/configure/editor-appearance/#toolbar_sticky)

### Sticky Toolbar
If you have a very long editor area then as you scroll the toolbar becomes sticky at the top of the screen so that it is always available

### Toolbar Drawer
This restricts the toolbar to only display one row (depending on the screen size) and if there are buttons that can not fit on the row then a `...` button appears which can be used to open and close the drawer to the other buttons.

To ensure that our `CMS Content` buttons are always displayed I moved them from the end to the beginning.

The Toolbar Drawer does not support the creation of multiple toolbars so to maintain backwards compatibility I merge the two toolbars into one.

Note: None of the presets were using the second toolbar

TinyMCE 5.2 will expand on this feature and allow greater customisation but they have not revealed how yet.

I have not made either of these as configurable settings - there are enough of those already.