[4.0] Remove Atum overrides
There is no need to have these overrides. We really shouldnt have to create overrides for the core templates and it is a maintenance headache.

This PR removes the overrides and puts the two tiny changes back into the core files.

For reference the changes are
1. Edit Article
Remove the box shadow around the wysiwyg area
2. Article Manager
Change the size of the text for the status text to small